Update the table with all symbols basic information

This function updates the 'all_symbols' table in the database with the latest stock symbols and their basic information from various indexes.

update_symbols_table(db_con, indexes = NULL, save_data = TRUE)

Arguments

db_con

A DBI database connection object.

indexes

A character vector specifying which indexes to filter by. Options include "ASX", "B3", "SP500", and "Benchmark". If NULL, returns all symbols.

Value

Nothing. The function updates the 'all_symbols' table in the database, informing the user of the update status.

Details

It fetches data from the ASX, B3, and S&P 500 indexes, combines them, and updates the database table accordingly.

Examples

if (FALSE) { # \dontrun{
# define the path and name of the database file
db_name <- "test_stock_db"
db_path <- file.path(tempdir(), db_name)
# create the database connection
db_con <- DBI::dbConnect(duckdb::duckdb(), dbdir = db_path, read_only = FALSE)
# Update the symbols table
update_symbols_table(indexes = c("ASX", "B3"), db_con = db_con)
DBI::dbDisconnect(db_con)
} # }